# server/docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "BidsHandler API Documentation (OpenAPIv3)",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "API server"
    }
  ],
...
  "paths": {
...
    "/api/v1/bids/": {
      "get": {
        "tags": [
          "Bids&Comments"
        ],
        "summary": "Get bids end-point.",
        "description": "This is the end-point to get bids.",
        "operationId": "getBids",
        "parameters": [
          {
            "$ref": "#/components/parameters/LimitInQuery"
          },
          {
            "$ref": "#/components/parameters/OffsetInQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseGetBidsSchema"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response400Schema"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response401Schema"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Bids&Comments"
        ],
        "summary": "Create bid end-point.",
        "description": "This is the end-point to create bid.",
        "operationId": "createBid",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response400Schema"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response401Schema"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCreateBidSchema"
              }
            }
          }
        }
      }
    },
    "/api/v1/bids/{id}": {
      "get": {
        "tags": [
          "Bids&Comments"
        ],
        "summary": "Get one bid end-point.",
        "description": "This is the end-point to get one bid.",
        "operationId": "getOneBid",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/IdInPath"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseGetOneBidSchema"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response400Schema"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response401Schema"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response404Schema"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Bids&Comments"
        ],
        "summary": "Process bid end-point.",
        "description": "This is the end-point to process bid with email notification to user.",
        "operationId": "processBid",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/IdInPath"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response400Schema"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response401Schema"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response404Schema"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestProcessBidSchema"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "LimitInQuery": {
        "in": "query",
        "name": "limit",
        "description": "Limit of items per page",
        "required": false,
        "default": 10,
        "schema": {
          "type": "integer",
          "format": "integer",
          "example": 10
        }
      },
      "OffsetInQuery": {
        "in": "query",
        "name": "page",
        "description": "Page number",
        "required": false,
        "default": 0,
        "schema": {
          "type": "integer",
          "format": "integer",
          "example": 0
        }
      },
      "IdInPath": {
        "in": "path",
        "name": "id",
        "description": "Request bid uuid",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid",
          "example": "c759d20a-d96c-40bc-9774-579d59744f24"
        }
      }
    },
    "schemas": {
...
      "RequestCreateBidSchema": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "Bid message",
            "description": "Bid message"
          }
        }
      },
      "RequestProcessBidSchema": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "example": "pending",
            "description": "Bid status"
          }
        }
      },
...
      "ResponseGetBidsSchema": {
        "type": "object",
        "required": [
          "bids",
          "count",
          "limit",
          "offset"
        ],
        "properties": {
          "count": {
            "type": "number",
            "example": 5,
            "description": "Bids count"
          },
          "limit": {
            "type": "number",
            "example": 10,
            "description": "Bids limit"
          },
          "offset": {
            "type": "number",
            "example": 0,
            "description": "Bids offset"
          },
          "bids": {
            "type": "array",
            "example": [
              {
                "id": "c759d20a-d96c-40bc-9774-579d59744f24",
                "status": "pending",
                "message": "Lorem ipsum dolor sit...",
                "authorId": "adb7247d-f032-4919-8da5-2d3987b046ee",
                "createdAt": "2025-04-13T20:39:11.834Z",
                "updatedAt": "2025-04-13T20:39:11.834Z",
                "deletedAt": null
              }
            ],
            "description": "Bids array; empty if no bids found"
          }
        }
      },
      "ResponseGetOneBidSchema": {
        "type": "object",
        "required": [
          "id",
          "status",
          "message",
          "authorId",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "c759d20a-d96c-40bc-9774-579d59744f24",
            "description": "Bid id (uuidv4)"
          },
          "status": {
            "type": "string",
            "example": "pending",
            "description": "Bid status"
          },
          "message": {
            "type": "string",
            "example": "Lorem ipsum dolor sit...",
            "description": "Bid message"
          },
          "authorId": {
            "type": "string",
            "example": "adb7247d-f032-4919-8da5-2d3987b046ee",
            "description": "Bid author id (uuidv4)"
          },
          "createdAt": {
            "type": "string",
            "example": "2025-04-13T20:39:11.834Z",
            "description": "Bid creation date"
          },
          "updatedAt": {
            "type": "string",
            "example": "2025-04-13T20:39:11.834Z",
            "description": "Bid last update date"
          },
          "deletedAt": {
            "type": "string",
            "example": "2025-04-13T20:39:11.834Z",
            "description": "Bid soft deletion date"
          }
        }
      },
      "Response400Schema": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "Bad Request",
            "description": "Error message"
          }
        }
      },
      "Response401Schema": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "Unauthorized!",
            "description": "Error message"
          }
        }
      },
      "Response404Schema": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "Not found!",
            "description": "Error message"
          }
        }
      }
    }
  }
}